adds optional revision in LookupPermissionSets
After the introduction of BreakingSchemaChange,
clients will be notified that an incompatible change has happened and
that they need rebuild their index. In order to do that, they need
to call LookupPermissionSets with the revision at which the breaking
event happened.

This introduces an optional revision in LookupPermissionSets to support
this use-case.